[Refactor] Refactor the image_* command flagging process
