+++ pkg/cmd/image/push.go
@@ -1,0 +1,217 @@
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package image
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+
+	"github.com/containerd/containerd/content"
+	"github.com/containerd/containerd/images"
+	"github.com/containerd/containerd/images/converter"
+	refdocker "github.com/containerd/containerd/reference/docker"
+	"github.com/containerd/containerd/remotes"
+	"github.com/containerd/nerdctl/pkg/api/types"
+	"github.com/containerd/nerdctl/pkg/clientutil"
+	"github.com/containerd/nerdctl/pkg/cosignutil"
+	"github.com/containerd/nerdctl/pkg/errutil"
+	"github.com/containerd/nerdctl/pkg/imgutil/dockerconfigresolver"
+	"github.com/containerd/nerdctl/pkg/imgutil/push"
+	"github.com/containerd/nerdctl/pkg/ipfs"
+	"github.com/containerd/nerdctl/pkg/platformutil"
+	"github.com/containerd/nerdctl/pkg/referenceutil"
+	"github.com/containerd/stargz-snapshotter/estargz"
+	"github.com/containerd/stargz-snapshotter/estargz/zstdchunked"
+	estargzconvert "github.com/containerd/stargz-snapshotter/nativeconverter/estargz"
+	"github.com/opencontainers/go-digest"
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+	"github.com/sirupsen/logrus"
+)
+
+func Push(ctx context.Context, rawRef string, options types.ImagePushCommandOptions, stdout io.Writer) error {
+	client, ctx, cancel, err := clientutil.NewClient(ctx, options.GOptions.Namespace, options.GOptions.Address)
+	if err != nil {
+		return err
+	}
+	defer cancel()
+
+	if scheme, ref, err := referenceutil.ParseIPFSRefWithScheme(rawRef); err == nil {
+		if scheme != "ipfs" {
+			return fmt.Errorf("ipfs scheme is only supported but got %q", scheme)
+		}
+		logrus.Infof("pushing image %q to IPFS", ref)
+
+		var ipfsPath *string
+		if options.IpfsAddress != "" {
+			dir, err := os.MkdirTemp("", "apidirtmp")
+			if err != nil {
+				return err
+			}
+			defer os.RemoveAll(dir)
+			if err := os.WriteFile(filepath.Join(dir, "api"), []byte(options.IpfsAddress), 0600); err != nil {
+				return err
+			}
+			ipfsPath = &dir
+		}
+
+		var layerConvert converter.ConvertFunc
+		if options.Estargz {
+			layerConvert = eStargzConvertFunc()
+		}
+		c, err := ipfs.Push(ctx, client, ref, layerConvert, options.AllPlatforms, options.Platforms, options.IpfsEnsureImage, ipfsPath)
+		if err != nil {
+			logrus.WithError(err).Warnf("ipfs push failed")
+			return err
+		}
+		fmt.Fprintln(stdout, c)
+		return nil
+	}
+
+	named, err := refdocker.ParseDockerRef(rawRef)
+	if err != nil {
+		return err
+	}
+	ref := named.String()
+	refDomain := refdocker.Domain(named)
+
+	platMC, err := platformutil.NewMatchComparer(options.AllPlatforms, options.Platforms)
+	if err != nil {
+		return err
+	}
+	pushRef := ref
+	if !options.AllPlatforms {
+		pushRef = ref + "-tmp-reduced-platform"
+		// Push fails with "400 Bad Request" when the manifest is multi-platform but we do not locally have multi-platform blobs.
+		// So we create a tmp reduced-platform image to avoid the error.
+		platImg, err := converter.Convert(ctx, client, pushRef, ref, converter.WithPlatform(platMC))
+		if err != nil {
+			if len(options.Platforms) == 0 {
+				return fmt.Errorf("failed to create a tmp single-platform image %q: %w", pushRef, err)
+			}
+			return fmt.Errorf("failed to create a tmp reduced-platform image %q (platform=%v): %w", pushRef, options.Platforms, err)
+		}
+		defer client.ImageService().Delete(ctx, platImg.Name, images.SynchronousDelete())
+		logrus.Infof("pushing as a reduced-platform image (%s, %s)", platImg.Target.MediaType, platImg.Target.Digest)
+	}
+
+	if options.Estargz {
+		pushRef = ref + "-tmp-esgz"
+		esgzImg, err := converter.Convert(ctx, client, pushRef, ref, converter.WithPlatform(platMC), converter.WithLayerConvertFunc(eStargzConvertFunc()))
+		if err != nil {
+			return fmt.Errorf("failed to convert to eStargz: %v", err)
+		}
+		defer client.ImageService().Delete(ctx, esgzImg.Name, images.SynchronousDelete())
+		logrus.Infof("pushing as an eStargz image (%s, %s)", esgzImg.Target.MediaType, esgzImg.Target.Digest)
+	}
+
+	pushFunc := func(r remotes.Resolver) error {
+		return push.Push(ctx, client, r, stdout, pushRef, ref, platMC, options.AllowNondistributableArtifacts)
+	}
+
+	var dOpts []dockerconfigresolver.Opt
+	if options.GOptions.InsecureRegistry {
+		logrus.Warnf("skipping verifying HTTPS certs for %q", refDomain)
+		dOpts = append(dOpts, dockerconfigresolver.WithSkipVerifyCerts(true))
+	}
+	dOpts = append(dOpts, dockerconfigresolver.WithHostsDirs(options.GOptions.HostsDir))
+	resolver, err := dockerconfigresolver.New(ctx, refDomain, dOpts...)
+	if err != nil {
+		return err
+	}
+	if err = pushFunc(resolver); err != nil {
+		// In some circumstance (e.g. people just use 80 port to support pure http), the error will contain message like "dial tcp <port>: connection refused"
+		if !errutil.IsErrHTTPResponseToHTTPSClient(err) && !errutil.IsErrConnectionRefused(err) {
+			return err
+		}
+		if options.GOptions.InsecureRegistry {
+			logrus.WithError(err).Warnf("server %q does not seem to support HTTPS, falling back to plain HTTP", refDomain)
+			dOpts = append(dOpts, dockerconfigresolver.WithPlainHTTP(true))
+			resolver, err = dockerconfigresolver.New(ctx, refDomain, dOpts...)
+			if err != nil {
+				return err
+			}
+			return pushFunc(resolver)
+		}
+		logrus.WithError(err).Errorf("server %q does not seem to support HTTPS", refDomain)
+		logrus.Info("Hint: you may want to try --insecure-registry to allow plain HTTP (if you are in a trusted network)")
+		return err
+	}
+
+	switch options.Sign {
+	case "cosign":
+
+		if !options.GOptions.Experimental {
+			return fmt.Errorf("cosign only work with enable experimental feature")
+		}
+
+		err = cosignutil.SignCosign(rawRef, options.CosignKey)
+		if err != nil {
+			return err
+		}
+	case "none":
+		logrus.Debugf("signing process skipped")
+	default:
+		return fmt.Errorf("no signers found: %s", options.Sign)
+
+	}
+
+	return nil
+}
+
+func eStargzConvertFunc() converter.ConvertFunc {
+	convertToESGZ := estargzconvert.LayerConvertFunc()
+	return func(ctx context.Context, cs content.Store, desc ocispec.Descriptor) (*ocispec.Descriptor, error) {
+		if isReusableESGZ(ctx, cs, desc) {
+			logrus.Infof("reusing estargz %s without conversion", desc.Digest)
+			return nil, nil
+		}
+		newDesc, err := convertToESGZ(ctx, cs, desc)
+		if err != nil {
+			return nil, err
+		}
+		logrus.Infof("converted %q to %s", desc.MediaType, newDesc.Digest)
+		return newDesc, err
+	}
+
+}
+
+func isReusableESGZ(ctx context.Context, cs content.Store, desc ocispec.Descriptor) bool {
+	dgstStr, ok := desc.Annotations[estargz.TOCJSONDigestAnnotation]
+	if !ok {
+		return false
+	}
+	tocdgst, err := digest.Parse(dgstStr)
+	if err != nil {
+		return false
+	}
+	ra, err := cs.ReaderAt(ctx, desc)
+	if err != nil {
+		return false
+	}
+	defer ra.Close()
+	r, err := estargz.Open(io.NewSectionReader(ra, 0, desc.Size), estargz.WithDecompressors(new(zstdchunked.Decompressor)))
+	if err != nil {
+		return false
+	}
+	if _, err := r.VerifyTOC(tocdgst); err != nil {
+		return false
+	}
+	return true
+}
